refactor: split CommandBuilder into partial classes by command group
