+++ src/officecli/CommandBuilder.GetQuery.cs
@@ -1,0 +1,103 @@
+// Copyright 2025 OfficeCli (officecli.ai)
+// SPDX-License-Identifier: Apache-2.0
+
+using System.CommandLine;
+using OfficeCli.Core;
+
+namespace OfficeCli;
+
+static partial class CommandBuilder
+{
+    private static Command BuildGetCommand(Option<bool> jsonOption)
+    {
+        var getFileArg = new Argument<FileInfo>("file") { Description = "Office document path (required even with open/close mode)" };
+        var pathArg = new Argument<string>("path") { Description = "DOM path (e.g. /body/p[1])" };
+        pathArg.DefaultValueFactory = _ => "/";
+        var depthOpt = new Option<int>("--depth") { Description = "Depth of child nodes to include" };
+        depthOpt.DefaultValueFactory = _ => 1;
+
+        var getCommand = new Command("get", "Get a document node by path");
+        getCommand.Add(getFileArg);
+        getCommand.Add(pathArg);
+        getCommand.Add(depthOpt);
+        getCommand.Add(jsonOption);
+
+        getCommand.SetAction(result => { var json = result.GetValue(jsonOption); return SafeRun(() =>
+        {
+            var file = result.GetValue(getFileArg)!;
+            var path = result.GetValue(pathArg)!;
+            var depth = result.GetValue(depthOpt);
+
+            if (TryResident(file.FullName, req =>
+            {
+                req.Command = "get";
+                req.Json = json;
+                req.Args["path"] = path;
+                req.Args["depth"] = depth.ToString();
+            }, json) is {} rc) return rc;
+
+            using var handler = DocumentHandlerFactory.Open(file.FullName);
+            var node = handler.Get(path, depth);
+            if (json)
+                Console.WriteLine(OutputFormatter.WrapEnvelope(
+                    OutputFormatter.FormatNode(node, OutputFormat.Json)));
+            else
+                Console.WriteLine(OutputFormatter.FormatNode(node, OutputFormat.Text));
+            return 0;
+        }, json); });
+
+        return getCommand;
+    }
+
+    private static Command BuildQueryCommand(Option<bool> jsonOption)
+    {
+        var queryFileArg = new Argument<FileInfo>("file") { Description = "Office document path (required even with open/close mode)" };
+        var selectorArg = new Argument<string>("selector") { Description = "CSS-like selector (e.g. paragraph[style=Normal] > run[font!=Arial])" };
+
+        var queryTextOpt = new Option<string?>("--text") { Description = "Filter results to elements containing this text (case-insensitive)" };
+
+        var queryCommand = new Command("query", "Query document elements with CSS-like selectors");
+        queryCommand.Add(queryFileArg);
+        queryCommand.Add(selectorArg);
+        queryCommand.Add(jsonOption);
+        queryCommand.Add(queryTextOpt);
+
+        queryCommand.SetAction(result => { var json = result.GetValue(jsonOption); return SafeRun(() =>
+        {
+            var file = result.GetValue(queryFileArg)!;
+            var selector = result.GetValue(selectorArg)!;
+            var textFilter = result.GetValue(queryTextOpt);
+
+            if (TryResident(file.FullName, req =>
+            {
+                req.Command = "query";
+                req.Json = json;
+                req.Args["selector"] = selector;
+                if (textFilter != null) req.Args["text"] = textFilter;
+            }, json) is {} rc) return rc;
+
+            var format = json ? OutputFormat.Json : OutputFormat.Text;
+
+            using var handler = DocumentHandlerFactory.Open(file.FullName);
+            var filters = OfficeCli.Core.AttributeFilter.Parse(selector);
+            var (results, warnings) = OfficeCli.Core.AttributeFilter.ApplyWithWarnings(handler.Query(selector), filters);
+            if (!string.IsNullOrEmpty(textFilter))
+                results = results.Where(n => n.Text != null && n.Text.Contains(textFilter, StringComparison.OrdinalIgnoreCase)).ToList();
+            if (json)
+            {
+                var cliWarnings = warnings.Select(w => new OfficeCli.Core.CliWarning { Message = w, Code = "filter_warning" }).ToList();
+                Console.WriteLine(OutputFormatter.WrapEnvelope(
+                    OutputFormatter.FormatNodes(results, OutputFormat.Json),
+                    cliWarnings.Count > 0 ? cliWarnings : null));
+            }
+            else
+            {
+                foreach (var w in warnings) Console.Error.WriteLine(w);
+                Console.WriteLine(OutputFormatter.FormatNodes(results, OutputFormat.Text));
+            }
+            return 0;
+        }, json); });
+
+        return queryCommand;
+    }
+}
